Rework oss-cad-suite install
The install script will determine the user with UID 1000. It will then
run second install script as the user, which will install
OSS CAD Suite in the user's home directory and update their .bashrc.

The two script approach prevents messing with UIDs/GIDs directly, which
was causing test issues in GitHub Action because the runner images run
with a UID of 1001 instead of 1000.